(feat:aluminium, aggregatesFromWarehouse, aggregatesWithOil) Adding entity
