--- server/services/reports/src/main/java/com/recykling/report/urtReport/entities/refrigeratorPowerCableWeights/repo/RefrigeratorPowerCableWeightsRepository.java
@@ -1,9 +1,0 @@
-package com.recykling.report.urtReport.entities.refrigeratorPowerCableWeights.repo;
-
-import com.recykling.report.urtReport.entities.refrigeratorPowerCableWeights.RefrigeratorPowerCableWeights;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RefrigeratorPowerCableWeightsRepository extends JpaRepository<RefrigeratorPowerCableWeights, Long> {
-}
